+++ spec/factories/event.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :event do
+    sequence(:title) { |n| "event#{n}" }
+    description { 'Test Event' }
+  end
+end